restructure tests and fix typo
